% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{correct_coding}
\alias{correct_coding}
\title{Correct coding}
\usage{
correct_coding(isomut, keep_uncorrected = FALSE, highlight = FALSE)
}
\arguments{
\item{isomut}{a data.frame of isomut data as generated by \code{read_isomut} and
annotated by \code{annotate_isomut}.}

\item{keep_uncorrected}{boolean. Adds a column for the uncorrected coding.}

\item{highlight}{boolean. Adds a column that indicates which variants have
corrected condons.}
}
\value{
a data.frame of the same format with corrected coding.
}
\description{
corrects the predicted AA if multiple variants are found in the same codon
}
