% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{format_indels}
\alias{format_indels}
\title{Format indels}
\usage{
format_indels(isomut, ref)
}
\arguments{
\item{isomut}{a data.frame of isomut data as generated by \code{read_isomut}.}

\item{ref}{either the name of a fasta file containing the reference genome or
an object of class \code{FaFile}.}
}
\value{
A data.frame with formatted sequence information.
}
\description{
Converts the sequence information about indels to a format that can be used
to predict the coding consequences.
}
