% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_variants}
\alias{plot_variants}
\title{Plot variants in genes or regions}
\usage{
plot_variants(
  isomut,
  regions,
  annotation = gtf_scer,
  ymax = NULL,
  fill = NULL,
  bar_width = NULL,
  flank = 50,
  showGenes = TRUE,
  fontsize = NULL,
  dy = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{isomut}{a \code{data.frame} or \code{data.table} of isomut data as generated by
\code{read_isomut}.}

\item{regions}{either a character vector of gene names or a \code{GRanges} object.}

\item{annotation}{path to annotation file in \code{gtf} or \code{gff} format or \code{GRanges}
object containing the annotations.}

\item{ymax}{maximum of variants to plot.}

\item{fill}{isomut column to use for fill color (default is \code{mut_freq}).}

\item{bar_width}{width of the bars (default 5).}

\item{flank}{additional space left and right of genes (in bp).}

\item{showGenes}{whether to show genes in the plot (only works with annotation).}

\item{fontsize}{font size for gene and chromosome names.}

\item{dy}{width of gene/chromosome bars.}

\item{filename}{optionally specify file name.}

\item{...}{additional parameters for \code{ggsave} (e.g. \code{width} and \code{height}).}
}
\description{
\code{plot_variants} plots the variants of an isomut file within a set of genes or
regions.
}
