% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read_isomut}
\alias{read_isomut}
\title{Read and preprocess an isomut file}
\usage{
read_isomut(
  file,
  minReads = 0L,
  minCoverage = 0,
  minMutFreq = 0,
  removePatterns = NULL,
  extraColumns = NULL
)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{minReads}{integer indicating the minimum number of reads
(default = 0L).}

\item{minCoverage}{double indicating the minimum coverage
(default = 0).}

\item{minMutFreq}{double indicating the minimum mutation frequency
(default = 0).}

\item{removePatterns}{character vector providing patterns in
sample_names that are to be removed.}

\item{extraColumns}{named list to generate extra columns. The list names are
the names of the new columns. The list elements can either be a vector of
length one (each row gets the same value for this column), a named vector
with names being a subset of the \code{sample_name} column in the original isomut
file, or a function that accepts those names as arguments.}
}
\value{
a data.frame containing the filtered and preprocessed data.
}
\description{
\code{read_isomut} reads an isomut file, adds a \code{read} column, optionally
filters according to given criteria and adds extra columns.
}
