% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read_isomut_from_dir}
\alias{read_isomut_from_dir}
\title{Read isomut data from directory}
\usage{
read_isomut_from_dir(
  dir,
  minReads = 0L,
  minCoverage = 0,
  minMutFreq = 0,
  minScore = 0,
  removePatterns = NULL,
  extraColumns = NULL,
  removeFeatures = TRUE,
  featureData = SGD_features,
  removeMito = TRUE,
  verbose = TRUE,
  asDataFrame = TRUE,
  annotate = TRUE,
  ...
)
}
\arguments{
\item{dir}{path to the directory containing isomut data.}

\item{minReads}{integer indicating the minimum number of reads
(default = 0L).}

\item{minCoverage}{double indicating the minimum coverage
(default = 0).}

\item{minMutFreq}{double indicating the minimum mutation frequency
(default = 0).}

\item{minScore}{double indicating the minimum isomut score
(default = 0).}

\item{removePatterns}{character vector providing patterns in
sample_names that are to be removed.}

\item{extraColumns}{named list to generate extra columns. The list names are
the names of the new columns. The list elements can either be a vector of
length one (each row gets the same value for this column), a named vector
with names being a subset of the \code{sample_name} column in the original isomut
file, or a function that accepts those names as arguments.}

\item{removeFeatures}{whether certain genomic features should be removed.}

\item{featureData}{\code{GRanges} that contains features to be removed.}

\item{removeMito}{whether variants in mitochondrial DNA should be removed.}

\item{verbose}{monitor progress.}

\item{asDataFrame}{whether to return a \code{data.frame}. If \code{FALSE}, a
\code{data.table} is returned.}

\item{annotate}{whether data should be annotated. For details see
\code{annotate_isomut}.}

\item{...}{further arguments for \code{annotate_isomut}.}
}
\value{
a \code{data.frame} or \code{data.table} containing the filtered and
preprocessed data.
}
\description{
Read isomut data from directory
}
