% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{remove_features}
\alias{remove_features}
\title{Remove features}
\usage{
remove_features(isomut, featureData = SGD_features)
}
\arguments{
\item{isomut}{a data.frame of isomut data as generated by \code{read_isomut}.}

\item{featureData}{a \code{GRanges} object that contains the features to be removed.}
}
\value{
a data.frame containing the filtered data.
}
\description{
Remove variants from isomut data if they match given genomic features
}
