% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{annotate_isomut}
\alias{annotate_isomut}
\title{Annotate isomut data}
\usage{
annotate_isomut(isomut, gtf_path, noDups = FALSE)
}
\arguments{
\item{isomut}{a data.frame of isomut data as generated by \code{read_isomut}
(minimally needs to contain \code{chr} and \code{pos} columns.}

\item{gtf_path}{path to annotation file in \code{gtf} or \code{gff} format.}

\item{noDups}{don't keep multiple annotations (e.g. when variant is inside
two genes on opposite strands). Default \code{FALSE}.}
}
\value{
a data.frame with additional columns regarding annotation.
}
\description{
\code{annotate_isomut} adds information about the location of variants (coding or
intergenic) and, for coding variants, about the affected gene(s) and the
position of the variant in the gene.
}
