% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{predict_coding}
\alias{predict_coding}
\title{Predict coding from variants}
\usage{
predict_coding(
  isomut,
  ref = system.file("extdata", "Saccharomyces_cerevisiae.R64-1-1.dna.toplevel.fa",
    package = "isomutR"),
  annotation = gtf_scer,
  txdb = NULL,
  format_indels = TRUE,
  correct_coding = FALSE
)
}
\arguments{
\item{isomut}{a data.frame of isomut data as generated by \code{read_isomut}.}

\item{ref}{either the name of a fasta file containing the reference genome or
an object of class \code{FaFile}.}

\item{annotation}{path to annotation file in \code{gtf} or \code{gff} format or \code{GRanges}
object containing the annotations.}

\item{txdb}{optionally provide \code{TxDb} object (should match \code{annotation}).}

\item{format_indels}{whether indels should be formatted (see \code{format_indels}
for details).}

\item{correct_coding}{whether coding be corrected if two variants hit the same
codon.}
}
\value{
a data.frame with additional columns regarding coding consequences.
}
\description{
predicts the consequences of variants in coding regions.
}
